# demo/src/main/java/com/example/demo/model/dto/SubscribeRequestDto.java
package com.example.demo.model.dto;

import lombok.AccessLevel;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.experimental.FieldDefaults;

@Data
@FieldDefaults(level = AccessLevel.PRIVATE)
@NoArgsConstructor
public class SubscribeRequestDto {
    String userId;
    String uri;
    double price;
}